conformity on isotp_user.h
To be able to compile with C++ compilers, it's common practice to offer extern "C" on a makro basis. All C-based system includes of Linux do so. I suggest to do it likewise here.

Another thing is to `#include <stdint.h>`. The problem is that some compilers often automatically export names defined in various headers or provided types before such standards were in place.
Details can be found here: https://stackoverflow.com/questions/11069108/uint32-t-does-not-name-a-type